+++ backend/connector-integration/src/connectors.rs
@@ -79,3 +79,9 @@ pub use self::aci::Aci;
 
 pub mod trustpay;
 pub use self::trustpay::Trustpay;
+
+pub mod stripe;
+pub use self::stripe::Stripe;
+
+pub mod cybersource;
+pub use self::cybersource::Cybersource;