+++ backend/connector-integration/src/types.rs
@@ -5,8 +5,8 @@ use interfaces::connector_types::BoxedConnector;
 
 use crate::connectors::{
     Aci, Adyen, Authorizedotnet, Bluecode, Braintree, Cashfree, Cashtocode, Checkout, Cryptopay,
-    Dlocal, Elavon, Fiserv, Fiuu, Helcim, Mifinity, Nexinets, Noon, Novalnet, Paytm, Payu, Phonepe,
-    Placetopay, Rapyd, Razorpay, RazorpayV2, Trustpay, Volt, Xendit,
+    Cybersource, Dlocal, Elavon, Fiserv, Fiuu, Helcim, Mifinity, Nexinets, Noon, Novalnet, Paytm,
+    Payu, Phonepe, Placetopay, Rapyd, Razorpay, RazorpayV2, Stripe, Trustpay, Volt, Xendit,
 };
 
 #[derive(Clone)]
@@ -56,6 +56,8 @@ impl<T: PaymentMethodDataTypes + Debug + Default + Send + Sync + 'static + serde
             ConnectorEnum::Rapyd => Box::new(Rapyd::new()),
             ConnectorEnum::Aci => Box::new(Aci::new()),
             ConnectorEnum::Trustpay => Box::new(Trustpay::new()),
+            ConnectorEnum::Stripe => Box::new(Stripe::new()),
+            ConnectorEnum::Cybersource => Box::new(Cybersource::new()),
         }
     }
 }